Update _utils.py
Hi there, and thank you for such an amazing package!

I've been trying to use the new 1.0 release to train DenseHMM models in a semi supervised fashion (that is, providing priors), and run across some errors that seem to be coming from the shape checks the .fit() method runs.

Upon inspecting the code, I saw the `_check_shapes()` function in _utils seems to be the culprit, since it checks that priors have the same shape as the input when they are passed. To my understanding, this should not be the case for HMMs, since priors should span the desired number of states, not the number of features available in the input data.

I modified the code in _check_shapes to make it work in my particular case. I'm submitting a pull request, and I'll be happy to take care of any other potential modifications needed to make this compatible with the package as a whole!

Best, and thank you once again for such terrific tool!
Lucas